mono-pic/src/processors/image_processor.py
"""
Image file processing for computer vision analysis
"""

import os
from typing import Dict, Any, List
from PIL import Image

class ImageProcessor:
    """Handles image file processing and computer vision analysis"""
    
    def __init__(self, config=None):
        self.config = config or {}
        self.supported_formats = ['.jpg', '.jpeg', '.png', '.bmp', '.gif']
        self.max_size = (1024, 1024)  # Max dimensions
        
    def validate_file(self, file_path: str) -> bool:
        """Validate if file is supported image format"""
        _, ext = os.path.splitext(file_path.lower())
        return ext in self.supported_formats
        
    def preprocess_image(self, file_path: str) -> str:
        """Preprocess image (resize, optimize)"""
        if not self.validate_file(file_path):
            raise ValueError(f"Unsupported image format: {file_path}")
            
        # TODO: Add image preprocessing logic
        # - Resize if too large
        # - Optimize file size
        # - Convert format if needed
        
        return file_path
        
    def extract_visual_features(self, file_path: str) -> Dict[str, Any]:
        """Extract visual features from image"""
        # TODO: Implement computer vision analysis
        # - Object detection
        # - Scene classification
        # - Color analysis
        # - Text extraction (OCR)
        
        return {
            'objects': [],
            'scene': '',
            'colors': [],
            'text': '',
            'faces': [],
            'emotions': []
        }
        
    def analyze_composition(self, file_path: str) -> Dict[str, Any]:
        """Analyze image composition and aesthetics"""
        # TODO: Implement composition analysis
        return {
            'rule_of_thirds': False,
            'brightness': 0.5,
            'contrast': 0.5,
            'saturation': 0.5,
            'focus_area': None
        }
        
    def process(self, file_path: str) -> Dict[str, Any]:
        """Main processing method"""
        processed_file = self.preprocess_image(file_path)
        visual_features = self.extract_visual_features(processed_file)
        composition = self.analyze_composition(processed_file)
        
        # Get basic image info
        with Image.open(file_path) as img:
            width, height = img.size
            
        return {
            'type': 'image',
            'source': file_path,
            'visual_features': visual_features,
            'composition': composition,
            'metadata': {
                'width': width,
                'height': height,
                'file_size': os.path.getsize(file_path),
                'format': os.path.splitext(file_path)[1]
            }
        }